Added Logging feature
git-svn-id: https://mediacenter-pc/svn/GPGPU/CUDA/RaytracingEngine/trunk@255 c4de0f11-4f5c-764a-a4aa-d9e147179871

+++ RayTracingEngineStub.cpp
@@ -216,7 +216,7 @@ extern "C" RAYTRACINGENGINE_API long RayTracer_SetMaterial(
    double reflection,
    double refraction,
    int    textured,
-   float  transparency,
+   double transparency,
    int    textureId,
    double specValue, 
    double specPower,
